remove fs-driver-android and update files
also making sure changes to file-api-driver-local is minimal
