+++ packages/app-mobile/package.json
@@ -51,7 +51,7 @@
     "react-native-popup-menu": "^0.10.0",
     "react-native-quick-actions": "^0.3.13",
     "react-native-rsa-native": "^2.0.4",
-    "react-native-saf-x": "^0.5.0",
+    "react-native-saf-x": "^0.5.2",
     "react-native-securerandom": "^1.0.0-rc.0",
     "react-native-share": "^7.2.1",
     "react-native-side-menu": "^1.1.3",
