+++ packages/app-mobile/utils/fs-driver-rn.ts
@@ -1,6 +1,13 @@
-import FsDriverBase from '@joplin/lib/fs-driver-base';
+import FsDriverBase, { ReadDirStatsOptions } from '@joplin/lib/fs-driver-base';
 const RNFetchBlob = require('react-native-blob-util').default;
 const RNFS = require('react-native-fs');
+import RNSAF, { Encoding, DocumentFileDetail } from 'react-native-saf-x';
+
+const ANDROID_URI_PREFIX = 'content://';
+
+function isScopedUri(path: string) {
+	return path.includes(ANDROID_URI_PREFIX);
+}
 
 export default class FsDriverRN extends FsDriverBase {
 	public appendFileSync() {
@@ -9,11 +16,17 @@ export default class FsDriverRN extends FsDriverBase {
 
 	// Encoding can be either "utf8" or "base64"
 	public appendFile(path: string, content: any, encoding = 'base64') {
+		if (isScopedUri(path)) {
+			return RNSAF.writeFile(path, content, { encoding: encoding as Encoding, append: true });
+		}
 		return RNFS.appendFile(path, content, encoding);
 	}
 
 	// Encoding can be either "utf8" or "base64"
 	public writeFile(path: string, content: any, encoding = 'base64') {
+		if (isScopedUri(path)) {
+			return RNSAF.writeFile(path, content, { encoding: encoding as Encoding });
+		}
 		// We need to use rn-fetch-blob here due to this bug:
 		// https://github.com/itinance/react-native-fs/issues/700
 		return RNFetchBlob.fs.writeFile(path, content, encoding);
@@ -26,52 +39,105 @@ export default class FsDriverRN extends FsDriverBase {
 
 	// Returns a format compatible with Node.js format
 	private rnfsStatToStd_(stat: any, path: string) {
+		let birthtime;
+		const mtime = stat.lastModified ? new Date(stat.lastModified) : stat.mtime;
+		if (stat.lastModified) {
+			birthtime = new Date(stat.lastModified);
+		} else if (stat.ctime) {
+			// Confusingly, "ctime" normally means "change time" but here it's used as "creation time". Also sometimes it is null
+			birthtime = stat.ctime;
+		} else {
+			birthtime = stat.mtime;
+		}
 		return {
-			birthtime: stat.ctime ? stat.ctime : stat.mtime, // Confusingly, "ctime" normally means "change time" but here it's used as "creation time". Also sometimes it is null
-			mtime: stat.mtime,
-			isDirectory: () => stat.isDirectory(),
+			birthtime,
+			mtime,
+			isDirectory: () => stat.type ? stat.type === 'directory' : stat.isDirectory(),
 			path: path,
 			size: stat.size,
 		};
+	}
+
+	public async isDirectory(path: string): Promise<boolean> {
+		return this.stat(path).then(result => result.isDirectory());
 	}
 
 	public async readDirStats(path: string, options: any = null) {
 		if (!options) options = {};
 		if (!('recursive' in options)) options.recursive = false;
 
-		let items = [];
+		const isScoped = isScopedUri(path);
+
+		let stats = [];
 		try {
-			items = await RNFS.readDir(path);
+			if (isScoped) {
+				stats = await RNSAF.listFiles(path);
+			} else {
+				stats = await RNFS.readDir(path);
+			}
 		} catch (error) {
 			throw new Error(`Could not read directory: ${path}: ${error.message}`);
 		}
 
 		let output: any[] = [];
-		for (let i = 0; i < items.length; i++) {
-			const item = items[i];
-			const relativePath = item.path.substr(path.length + 1);
-			output.push(this.rnfsStatToStd_(item, relativePath));
+		for (let i = 0; i < stats.length; i++) {
+			const stat = stats[i];
+			const relativePath = (isScoped ? stat.uri : stat.path).substr(path.length + 1);
+			output.push(this.rnfsStatToStd_(stat, relativePath));
 
-			output = await this.readDirStatsHandleRecursion_(path, item, output, options);
+			if (isScoped) {
+				output = await this.readUriDirStatsHandleRecursion_(stat, output, options);
+			} else {
+				output = await this.readDirStatsHandleRecursion_(path, stat, output, options);
+			}
+		}
+		return output;
+	}
+
+
+	protected async readUriDirStatsHandleRecursion_(stat: DocumentFileDetail, output: DocumentFileDetail[], options: ReadDirStatsOptions) {
+		if (options.recursive && stat.type === 'directory') {
+			const subStats = await this.readDirStats(stat.uri, options);
+			for (let j = 0; j < subStats.length; j++) {
+				const subStat = subStats[j];
+				output.push(subStat);
+			}
 		}
 		return output;
 	}
 
 	public async move(source: string, dest: string) {
+		if (isScopedUri(source) && isScopedUri(dest)) {
+			await RNSAF.moveFile(source, dest, { replaceIfDestinationExists: true });
+		} else if (isScopedUri(source) || isScopedUri(dest)) {
+			throw new Error('Move between different storage types not supported');
+		}
 		return RNFS.moveFile(source, dest);
 	}
 
 	public async exists(path: string) {
+		if (isScopedUri(path)) {
+			return RNSAF.exists(path);
+		}
 		return RNFS.exists(path);
 	}
 
 	public async mkdir(path: string) {
+		if (isScopedUri(path)) {
+			await RNSAF.mkdir(path);
+			return;
+		}
 		return RNFS.mkdir(path);
 	}
 
 	public async stat(path: string) {
 		try {
-			const r = await RNFS.stat(path);
+			let r;
+			if (isScopedUri(path)) {
+				r = await RNSAF.stat(path);
+			} else {
+				r = await RNFS.stat(path);
+			}
 			return this.rnfsStatToStd_(r, path);
 		} catch (error) {
 			if (error && ((error.message && error.message.indexOf('exist') >= 0) || error.code === 'ENOENT')) {
@@ -93,6 +159,9 @@ export default class FsDriverRN extends FsDriverBase {
 	}
 
 	public async open(path: string, mode: number) {
+		if (isScopedUri(path)) {
+			throw new Error('open() not implemented in FsDriverAndroid');
+		}
 		// Note: RNFS.read() doesn't provide any way to know if the end of file has been reached.
 		// So instead we stat the file here and use stat.size to manually check for end of file.
 		// Bug: https://github.com/itinance/react-native-fs/issues/342
@@ -112,6 +181,9 @@ export default class FsDriverRN extends FsDriverBase {
 
 	public readFile(path: string, encoding = 'utf8') {
 		if (encoding === 'Buffer') throw new Error('Raw buffer output not supported for FsDriverRN.readFile');
+		if (isScopedUri(path)) {
+			return RNSAF.readFile(path, { encoding: encoding as Encoding });
+		}
 		return RNFS.readFile(path, encoding);
 	}
 
@@ -119,6 +191,12 @@ export default class FsDriverRN extends FsDriverBase {
 	public async copy(source: string, dest: string) {
 		let retry = false;
 		try {
+			if (isScopedUri(source) && isScopedUri(dest)) {
+				await RNSAF.copyFile(source, dest, { replaceIfDestinationExists: true });
+				return;
+			} else if (isScopedUri(source) || isScopedUri(dest)) {
+				throw new Error('Move between different storage types not supported');
+			}
 			await RNFS.copyFile(source, dest);
 		} catch (error) {
 			// On iOS it will throw an error if the file already exist
@@ -131,6 +209,10 @@ export default class FsDriverRN extends FsDriverBase {
 
 	public async unlink(path: string) {
 		try {
+			if (isScopedUri(path)) {
+				await RNSAF.unlink(path);
+				return;
+			}
 			await RNFS.unlink(path);
 		} catch (error) {
 			if (error && ((error.message && error.message.indexOf('exist') >= 0) || error.code === 'ENOENT')) {
